Incorporated code review comments
