+++ src/test/java/uk/gov/hmcts/reform/prl/controllers/DraftAnOrderControllerTest.java
@@ -3,11 +3,11 @@ package uk.gov.hmcts.reform.prl.controllers;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.Assert;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
 import org.springframework.context.annotation.PropertySource;
@@ -184,7 +184,6 @@ public class DraftAnOrderControllerTest {
     }
 
     @Test
-    @Ignore
     public void testGenerateDoc() throws Exception {
         CaseData caseData = CaseData.builder()
             .id(123L)
@@ -203,7 +202,7 @@ public class DraftAnOrderControllerTest {
                              .data(stringObjectMap)
                              .build())
             .build();
-        when(draftAnOrderService.generateDocument(callbackRequest, caseData)).thenReturn(caseData);
+        when(draftAnOrderService.generateOrderDocument(Mockito.anyString(), Mockito.any(CallbackRequest.class))).thenReturn(stringObjectMap);
         Map<String, Object> caseDataUpdated = callbackRequest.getCaseDetails().getData();
         caseDataUpdated.putAll(manageOrderService.getCaseData("test token", caseData, CreateSelectOrderOptionsEnum.blankOrderOrDirections));
 
@@ -211,7 +210,6 @@ public class DraftAnOrderControllerTest {
     }
 
     @Test
-    @Ignore
     public void testPrepareDraftOrderCollection() throws Exception {
         CaseData caseData = CaseData.builder()
             .id(123L)
@@ -230,7 +228,7 @@ public class DraftAnOrderControllerTest {
                              .data(stringObjectMap)
                              .build())
             .build();
-        when(draftAnOrderService.generateDocument(callbackRequest, caseData)).thenReturn(caseData);
+        when(draftAnOrderService.prepareDraftOrderCollection(Mockito.anyString(), Mockito.any(CallbackRequest.class))).thenReturn(stringObjectMap);
         Map<String, Object> caseDataUpdated = callbackRequest.getCaseDetails().getData();
         caseDataUpdated.putAll(manageOrderService.getCaseData("test token", caseData, CreateSelectOrderOptionsEnum.blankOrderOrDirections));
