+++ src/test/java/uk/gov/hmcts/reform/prl/services/FL401SubmitApplicationServiceTest.java
@@ -49,6 +49,8 @@ import uk.gov.hmcts.reform.prl.services.tab.alltabs.AllTabServiceImpl;
 import uk.gov.hmcts.reform.prl.services.validators.FL401StatementOfTruthAndSubmitChecker;
 import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
@@ -963,5 +965,94 @@ public class FL401SubmitApplicationServiceTest {
         assertTrue(response.containsKey(DOCUMENT_FIELD_FINAL));
         assertTrue(response.containsKey(DOCUMENT_FIELD_C8_WELSH));
         assertTrue(response.containsKey(DOCUMENT_FIELD_FINAL_WELSH));
+    }
+
+    @Test
+    public void testFl401SendApplicationNotification() throws Exception {
+
+        PartyDetails fl401Applicant = PartyDetails.builder()
+            .canYouProvideEmailAddress(YesOrNo.No)
+            .isAddressConfidential(YesOrNo.No)
+            .isPhoneNumberConfidential(YesOrNo.No)
+            .build();
+
+        String applicantNames = "TestFirst TestLast";
+
+        String isConfidential = "No";
+        if (fl401Applicant.getCanYouProvideEmailAddress().equals(YesOrNo.Yes)
+            || (fl401Applicant.getIsEmailAddressConfidential() != null
+            && fl401Applicant.getIsEmailAddressConfidential().equals(YesOrNo.Yes))
+            || (fl401Applicant.hasConfidentialInfo())) {
+            isConfidential = "Yes";
+        }
+
+        CaseData caseData = CaseData.builder()
+            .id(12345L)
+            .applicantCaseName("TestCaseName")
+            .applicantsFL401(fl401Applicant)
+            .courtEmailAddress("localcourt@test.com")
+            .isNotificationSent("Yes")
+            .submitCountyCourtSelection(dynamicList)
+            .build();
+
+        LocalDate issueDate = LocalDate.now();
+        DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd-MM-yyyy");
+        Map<String, Object> stringObjectMap = new HashMap<>();
+
+        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
+        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
+        when(allTabsService.getAllTabsFields(any(CaseData.class))).thenReturn(stringObjectMap);
+
+        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
+            .CallbackRequest.builder().caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder().id(1L)
+                                                       .data(stringObjectMap).build()).build();
+
+        when(userService.getUserDetails(authToken)).thenReturn(userDetails);
+
+        fl401SubmitApplicationService.fl401SendApplicationNotification(authToken, callbackRequest);
+        verify(caseWorkerEmailService, times(0))
+            .sendEmailToFl401LocalCourt(callbackRequest.getCaseDetails(), caseData.getCourtEmailAddress());
+        verify(solicitorEmailService, times(1)).sendEmailToFl401Solicitor(
+            callbackRequest.getCaseDetails(),
+            userDetails
+        );
+    }
+
+    @Test
+    public void testFl401SendApplicationNotificationFailure() throws Exception {
+
+        PartyDetails fl401Applicant = PartyDetails.builder()
+            .canYouProvideEmailAddress(YesOrNo.No)
+            .isAddressConfidential(YesOrNo.No)
+            .isPhoneNumberConfidential(YesOrNo.No)
+            .build();
+
+        CaseData caseData = CaseData.builder()
+            .id(12345L)
+            .applicantCaseName("TestCaseName")
+            .applicantsFL401(fl401Applicant)
+            .courtEmailAddress("localcourt@test.com")
+            .isNotificationSent("No")
+            .build();
+
+        Map<String, Object> stringObjectMap = new HashMap<>();
+
+        when(objectMapper.convertValue(stringObjectMap, CaseData.class)).thenReturn(caseData);
+        when(userService.getUserDetails(Mockito.anyString())).thenReturn(userDetails);
+        when(allTabsService.getAllTabsFields(any(CaseData.class))).thenReturn(stringObjectMap);
+
+        uk.gov.hmcts.reform.ccd.client.model.CallbackRequest callbackRequest = uk.gov.hmcts.reform.ccd.client.model
+            .CallbackRequest.builder().caseDetails(uk.gov.hmcts.reform.ccd.client.model.CaseDetails.builder().id(1L)
+                                                       .data(stringObjectMap).build()).build();
+
+        when(userService.getUserDetails(authToken)).thenReturn(userDetails);
+
+        fl401SubmitApplicationService.fl401SendApplicationNotification(authToken, callbackRequest);
+        verify(caseWorkerEmailService, times(0))
+            .sendEmailToFl401LocalCourt(callbackRequest.getCaseDetails(), caseData.getCourtEmailAddress());
+        verify(solicitorEmailService, times(1)).sendEmailToFl401Solicitor(
+            callbackRequest.getCaseDetails(),
+            userDetails
+        );
     }
 }
